implemented react navigation for routing among components - able to route to all components
