+++ package.json
@@ -14,6 +14,7 @@
     "@material-ui/core": "^4.12.3",
     "@material-ui/icons": "^4.11.2",
     "@react-navigation/native": "^6.0.7",
+    "@react-navigation/native-stack": "^6.3.0",
     "crypto": "^1.0.1",
     "dotenv": "^14.3.2",
     "expo": "^44.0.0",
@@ -31,7 +32,7 @@
     "react-dom": "17.0.1",
     "react-native": "0.64.3",
     "react-native-dotenv": "^3.3.1",
-    "react-native-router-flux": "^4.3.1",
+    "react-native-router-flux": "^3.39.1",
     "react-native-web": "0.17.1",
     "react-navigation": "^4.4.4",
     "react-redux": "^7.2.6",
